requirements.txt
aestate-json==1.0.3
ahk==1.8.4
ahk-json==0.1.0rc2
allure-pytest==2.15.3
allure-python-commons==2.15.3
attrs==25.4.0
autopep8==2.3.2
black==26.1.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
configparser==7.2.0
flake8==7.3.0
greenlet==3.3.1
h11==0.16.0
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mccabe==0.7.0
mypy_extensions==1.1.0
outcome==1.3.0.post0
package_name==0.1
packaging==25.0
pathspec==1.0.3
platformdirs==4.5.1
pluggy==1.6.0
psycopg2-binary==2.9.11
pycodestyle==2.14.0
pycparser==2.23
pyflakes==3.4.0
Pygments==2.19.2
PySocks==1.7.1
pytest==9.0.2
python-dotenv==1.2.1
python-rd-python-gitlab==0.0.1
pytokens==0.4.0
requests==2.32.5
selenium==4.39.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.46
trio==0.32.0
trio-websocket==0.12.2
typing_extensions==4.15.0
urllib3==2.6.2
webdriver-manager==4.0.2
websocket-client==1.9.0
wsproto==1.3.2
